Make the BlockChecks work
As the BO3 creation logic was getting too big for one method, it's moved
to it's own class, BO3Creator. BO3CenterCreator is renamed to
BO3ClickHandler, because it now also selects the BlockChecks.

One issue I had is what to do with the block data in the block checks.
Block data is used for both rotations (for example stairs) and
submatierials (for example colored wool). By calling Terrain Control's
block rotator, I check whether the data is used for rotations. I let the
BlockCheck check for block data when it's used for submaterials, and not
for rotations.

The tree setting will now be set automatically. If the BO3 doesn't
contain leaves, the tree setting will be set to false.